Changed world tests to be more granular with versions
